[desktop integration] Add AppData file
AppData is a standard to describe GUI applications. In particular, it
allows these applications to be searched for and installed through
software centers on Fedora, Ubuntu, etc.

This commit embeds a `.appdata.xml` file for Avidemux, so my grandmother
can install it without having to open a terminal.

For further information on AppData see:
https://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html#sect-Quickstart-DesktopApps

Related to: https://github.com/rpmfusion/avidemux/pull/6