[core/cosmetics] Remove redundant path separators
Both canonize() and ADM_getInstallRelativePath() add a trailing directory separator
to the path. It is safe to remove trailing slashes after i18n in ADM_getI8NDir()
for Linux and the redundant path separator after pluginDir in ADM_getPluginDir().

The same applies to ADM_getI8NDir() in ADM_folder_mac.cpp, but can't test on Mac.