[core/muxer] Assume that MPEG is always MPEG-2
This is actually what libavformat is doing, which also specifies mpg1 as the primary tag for mpeg1video. Disadvantages from mislabeling almost extinct MPEG-1 as MPEG-2, if any, seem to be smaller than vice versa (e.g. mpv currently refusing to enable hwaccel if it encouters MKV with mislabeled MPEG-2).