[fadeToBlack] Remove overzealous bounds checks, the one in configure() is enough
The totalDuration reported by the bridge is *not* necessarily the total duration of the video, duh. This was a regression from b20b248, it resulted in configuration being reset when the difference between the end and the start time of the bridge was less than the start time of the filter.