[osxInstaller] Fix bundle creation with VapourSynth demuxer plugin enabled
VapourSynth demuxer pulls in libvapoursynth-script which in turn pulls in the huge Python framework which can't be relinked and we end up using external libs anyway. We relink the VapourSynth demuxer dylib back against external libvapoursynth and libvapoursynth-script and remove Python. The demuxer requires the vapoursynth package from Homebrew installed with default prefix to work.