[editor,muxerGate/videoCopyFromAnnexB] Try to cope better with H.264 SPS changing midstream in copy mode
When checking viability of cut points in H.264 streams, we need to make sure we use the right SPS info to extract POC and frame type from slice headers, else users face bogus warnings or are not warned when due.

This patch adds checks for in-band SPS not matching the global one and updates SPS info when necessary. It also changes the behavior of the AnnexB to mp4 converter to create H.264 and HEVC extradata from parameter sets directly at the start of the range we save instead of the first frame.