[demuxers/Matroska] Try to fix generated dts being incompatible with advertised time base resulting in failure to save as mp4
Always generate dts from sorted list of pts when B-frames are present, use 1/90000 timebase as fallback when pts cannot be represented sufficiently close as a multiple of a timebase matching minimum delta or a standard frame rate.

Do not penalize candidate frame rates for holes in pts. This is needed to disentangle frame rate and time base but may regress protection against double frame rate for now, to be added later.

Use RMSD for deviation calculation.

Keep pts for VC-1 as seemingly the lesser evil, wrong frame order should be caught by checks in the editor.