[audioFilter/copy,gui_savenew] Update properties of the audio stream for given start time when saving audio track in copy mode
If channel layout switches midstream e.g. from stereo to 5.1 surround and we save a portion of the video starting after the switch in copy mode for audio, this change fixes the issue that the output file falsely claims to contain a stereo audio track instead of a surround one.