[audioDecoders/ac3] Output silence with initial number of channels and sampling frequency if audio properties change midstream and it is not a stereo to 5.1 switch where a52dec performs downmixing
The client is supposed to call reconfigureCompleted() when ready to accept output with the new number of channels and sampling frequency. This patch allows to start playback of an AC3 audio track after the switch and to get the correct output. It doesn't help when videos with mismatching AC3 audio tracks are appended as it requires the decoder instance to stay the same.

The case of sampling frequency changing on the fly is completely untested and therefore probably broken.