[x265] Add option for intra in B frames
This is turned on in presets medium and above and off for presets faster
than medium.

Except it's off in the slow preset, which is probably a bug in x265.
There's an issue for it.

Making this an option allows duplicating faster and below presets via
advanced config.  And allows fixing the probable slow reset bug to turn
it on.

It's combined in the same UI row with weighted prediction in B-frames.