[x265] Add options for reference frame limit
This is the limit-refs option which is changed at the slow and above
presets to stop limiting references frames as much.

Adding it necessary to allow the slow preset to be duplicated with
advanced configuration.

It's added to the UI on the same line as number of ref frames:

Maximum Reference Frames: 3↕  Limit References by: [] CU [] Depth

When the max ref frames is 1, there's no effect to limit further, so the
checkboxes are disabled.  This is similar to how the two fastest
presets, which have max ref 1, have a limit-refs value of 0 (no limits),
despite that being the slower setting, since the limits have no effect.