[x265] Fix layout of Frame tab B-frames group
The labels and the spinners weren't lining up correctly.  It was because
there were four rows of controls but the grid layout had five rows.  A
different item in each column had a span of two rows to fill the grid.

Fix it to four rows with all items one row high.

Reorder the items in the markup for this grid to be in left to right,
top to bottom order, which is the normal order and makes the markup
easier to read.  They were in some random order before.