[x265] Add limit-modes option
This is turned on in "slow", when rect is enabled, and then turned off
in "veryslow".  It limits the search for additional motion partitions,
which are enabled with rect and amp.  So it only applies when at least
rect is enabled.  That's why it not on below the slow preset.

The UI will automatically disable the checkbox when rect is not enabled.