[x265] Refactor duplicated combo box code
Store data with each item in a combo box.  This avoids lots of code to
loop over lists searching for what item is in the box.

Make some template functions to fill the comboboxes with data.  I could
make them better if C++11 was allowed.

This fixes a bug where the poolthreads box didn't account for adding
"none" to the combobox when it wasn't in the list of possible values.
Using the item data makes this easier to deal with.

Another bug where none could be translated is fixed too.

Also use a translated "none" for the preset value.