[x265] Add ability to include color information
Add the ability to select the color primaries, color matrix, and transfer
characteristics.  This was already in the Output 2 tab, but was disabled and
never completed.

This will only be added to the bitstream if any of options are changed
from "Unknown".

I've tried to sort the available options in a useful and consistent way
and also indicate what kind of material will use them.  "Unknown" is
translatable, but the other items aren't.  I omitted the reserved codes.

The UI is reorganized significantly for this.  Create new tab "Display"
and put the PAR and color options here.  Then combine what was left of
the Output 1 and Output 2 tabs into one Output tab.  That tab remains
disabled.  The Video Format combo box also remains disabled on the
Display tab.

I've tested with ffprobe -show_stream and it finds the indicated color
information.