[x265] Implement pixel aspect ratio dialog
The existing code didn't work.

Fix it so that selecting unspecified will leave the PAR unspecified,
predefined will select one of the H.265 predefined SAR codes, and custom
will select an extended SAR.

The predefined aspect ratios are now the ones defined in H.265.  The PAR
values previously in the combo box are not in the standard and they are
(arguably) incorrect.