[demuxers/MpegTS] Fix for 60000/1001 framerate
The ts.idx2 file stores framerate as fps * 1000, instead of a rational
number with an integer number and denominator, which would be much
better.

It must then convert that back into a rational so it can put inside a
fake AVI header.

It does not use a closest rational approximation algorithm but just hard
codes a few different common frame rates.  And the standard ATSC rate of
60000/1001 was not in the list.  So add it.