[fadeTo] Discard the image we fade to / from on seek
When we seek to time within the range of this filter, we likely miss the
frame we actually want. This is usually not a problem for saving video
as the filter chain will be rebuilt and the wrong image cleared, but
it poses a problem when working in the filter manager.

This patch addresses only the issue of the wrong image persisting after
seek.