[crop] Use QSettings to store rubber band visibility
In the configuration dialog of the crop filter, the rubber band does not
need to be visible to use it once users have understood the functionality.
This patch is based on an assumption that users are unlikely to expect
rubber band visibility to change when they configure this filter next
time. Therefore, filter parameters were not the right place to store this
property as it needs to be per class rather than per instance.