[Qt/mainfilter] Fix crash on closing filter preview and filter manager in quick succession
Hitting the ESC key twice to close both filter preview and filter manager
could result in the second rejected() signal received by the preview
which was already removed from widget stack and subsequent assert
failure in qtUnregisterDialog(). This was especially easy to reproduce
during playback.

This patch avoids the crash by suppressing subsequent signals from filter
preview dialog. If ESC is pressed the second time too soon, the keypress
event is ignored which is not perfect but better than a crash.