[muxerGate] Exclude end of selection from range when saving video
The policy to include the end, while being intuitive, was not just
making it a pain to split a range using scripting but proved to be not
viable in the standard scenario of saving a part of an open GOP type
H.264 or HEVC stream in copy mode with the end of selection placed on a
keyframe as such keyframe may be not standalone decodable.

The problem of the last frame of video dropped unless the markers are
reset needs a different solution.