[editor] Fix previousPicture() skipping segment not containing a keyframe, always check success of linear to time in reference conversion
We were using an unsigned integer data type in nextPicture() to compare
timing of a cached picture with segment start. This resulted in cached
pictures from before the segment start getting unjustly discarded,
breaking navigation to the previous picture at segment boundary when the
previous segment didn't contain a keyframe.