[Qt/menu] Allow to explicitly disable linked controls
While generally redundant as we just need to enable linked controls for
menu entries where we want them to be enabled and they will be implicitly
disabled for all the others, ignoring the value of the second argument of

diaElemMenuDynamic::link(diaMenuEntryDynamic *, uint32_t, diaElem *);

may be perceived as a bug. Change the behavior to match expectations.