[ffNvEnc] Add maximum number of reference frames to encoder configuration
We have no means to let user know if the hardware does not support custom
number of reference frames, so with B-frames enabled, encoder setup may
fail when max refs is set to a non-zero value. Nevertheless, the resulting
video stream will have the necessary number of ref frames. Therefore, the
default is 0 which means automatic.