[GUI_vdpauRender] Try to recover from width and height mismatch between mixer and input surface
Some codecs supported by hw decoders including but probably not limited to
HEVC use coded width and height rounded up to a multiple of 64 rather than
of 16 we assume, which can result in failure to display the picture.

Query the actual size of the hw image passed to the renderer and re-create
mixer in case of mismatch.