[file_qt4] Fix filename collision detection, broken on Windows
Qt uses forward slashes as directory separators on Windows as well, but
the path to the last read file (i.e. to the currently loaded file) may be
stored with backslashes. In this case we failed to detect name collision and
didn't automatically append _edit to the basename.

Canonicalize paths before comparing to make filename collision detection
more robust.