[cmake] Don't set CMAKE_EDIT_COMMAND
CMAKE_EDIT_COMMAND is mean to be an internal variable, used only
for implementing the 'make edit_cache' target of Makefile
Generators. It's supposed to contain the path to a CMake editor
like ccmake or cmake-gui, and will be called with arguments not
accepted by vim, mcedit, or any other generic editor.

https://cmake.org/cmake/help/latest/variable/CMAKE_EDIT_COMMAND.html
Signed-off-by: FeRD (Frank Dana) <ferdnyc@gmail.com>