[Q_gui2] Add 'Reset Marker A/B' buttons to 'Marker A/B' fields
With being unable to seek beyond the last frame, the reset marker B
button is useful to put the marker B at total time position (beyond
the last frame).

To put the marker B at total time position (beyond the last frame),
set the marker A, then reset the marker B.

The reset marker A button resets to position 0, the B button to the
position just after the last frame (aka infinite).

The presence of the reset marker A/B buttons also give a hint about
when the markers are unset, that is when a button is grayed out.