[Q_gui2] Fix to manage signals when time fields are readonly
This is a workaround about editingFinished events.  Even when time
fields are set to readonly, the signal triggers just one time when
the field is clicked and then the focus is changed, with Tab or by
clicking another element.

This could be a better solution than disabling the time fields, if
there are action buttons that should be kept enabled in a field.