[Q_gui2] Fix markers A/B time fields with obsolete text
Editing the markers A/B fields directly, by changing the text, may
result in a computed frame real position not diffrent from before.

Different timings may resolve to the same frame.  When an A/B text
refresh is not performed anyway, the edited value may be obsolete.

Editing A/B directly is more flexible than the 'Go to Time' dialog
but, as said before, the difference is that the former changes the
field text before the frame computation is done, the latter intead
changes the text after, and only if the computed time differs from
the currently one set.