[Q_gui2] Fix total time field with obsolete text
Confirming the total time edit should be followed by a refresh.

The total time field can be edited to prepare a cut selection from the
marker A to B starting from the end of the video.  This feature can be
used to set a cut to reduce the video length of a specific amount.

After refreshing the total time field, the video length specified will
be indicated by the marker A time.