audioFilterConfigure: safeguard against crash from bogus tempo ratio hint on opening audio filters config dialog
The assumption that the ratio between video durations at the start and at
the end of filter chain may be used to preset audio stretcher on opening
audio filters config dialog is valid for the single scenario of changeFps
video filter being applied to a video with an internal audio track as the
only video filter modifying duration.

With stillimage video filter the calculated tempo ratio will be bogus and
easily vastly out of range.

This patch is a stop-gap measure to avoid loss of functionality.