Add a Desktop entry for the job queue runner
On Linux, we currently can only launch the GUI application to run queued
jobs via the command line. This is not very intuitive and makes it hard
to use the queue feature. Adding a desktop entry for the runner should
make it more easily discoverable how to run the queued jobs.